:sparkles: Ask user for confirmation before delete

+++ src/infrastructure/remove_duplicates.py
@@ -3,7 +3,7 @@ import matplotlib.pyplot as plt
 import os
 from PIL import Image
 
-os.chdir(r'/Users/cyrillemaire/Documents/Yotta/Project/Project_2/dup_test/paris')
+os.chdir(r'/Users/cyrillemaire/Documents/Yotta/Project/Project_2/pictures/Venise')
 print(os.getcwd())
 
 files_list = os.listdir('.')
@@ -16,7 +16,6 @@ hash_keys = dict()
 for index, filename in enumerate(os.listdir('.')):
     if not filename.startswith('.'):
         if os.path.isfile(filename):
-            print(os.getcwd())
             img = Image.open(filename)
             # reduce image:
             img = img.resize((10, 10), Image.ANTIALIAS)
@@ -28,13 +27,13 @@ for index, filename in enumerate(os.listdir('.')):
             # Compute hash in base 16:
             bits = "".join(['1' if (px >= avg_pixel) else '0' for px in pixel_data])
             filehash = str(hex(int(bits, 2)))[2:][::-1].upper()
-            #filehash = hashlib.md5(f.read()).hexdigest()
 
             if filehash not in hash_keys:
                 hash_keys[filehash] = index
             else:
                 duplicates.append((index, hash_keys[filehash]))
 
+print(len(duplicates), "duplicates have been found")
 print(duplicates)
 
 for file_indexes in duplicates[:30]:
@@ -42,13 +41,17 @@ for file_indexes in duplicates[:30]:
         plt.subplot(121), plt.imshow(imread(files_list[file_indexes[1]]))
         plt.title(file_indexes[1]), plt.xticks([]), plt.yticks([])
 
-        plt.subplot(122),plt.imshow(imread(files_list[file_indexes[0]]))
+        plt.subplot(122), plt.imshow(imread(files_list[file_indexes[0]]))
         plt.title(str(file_indexes[0]) + 'duplicate'), plt.xticks([]), plt.yticks([])
         plt.show()
 
     except OSError as e:
         continue
 
-for index in duplicates:
-    os.remove(files_list[index[0]])
+ans = input('Do you want to delete the duplicates? [y/n]')
 
+if ans == 'y' or 'yes':
+    for index in duplicates:
+        os.remove(files_list[index[0]])
+else:
+    print('Nothing deleted, please review pictures manually or run the script again')
